error handling

+++ project_1/command.c
@@ -21,22 +21,17 @@ void listDir() {
 void showCurrentDir(){
     char* buffer;
     char *cwd = getcwd(buffer,256);
-    
-    if (cwd == NULL) {
-        perror("getcwd");
-        return EXIT_FAILURE;
-    }
-
-    write(1,cwd,strlen(cwd));
-    
+    write(1,cwd,strlen(cwd)); 
 }
 
 
 void makeDir(char *dirName){
     int result = mkdir(dirName, 0777);
 
-    if(result = 1){
-        write(1,"error creating directory",strlen("error creating directory"));
+    if(result == 1){
+        char error[256];
+        snprintf(error, sizeof(error), "mkdir: cannot create directory '%s': File exists\n", dirName);
+        write(1,error,strlen(error));
     }
 }
 
@@ -45,14 +40,29 @@ void changeDir(char *dirName){
     int result = chdir(dirName);
 
     if (result != 0){
-        write(1,"cd: no such file or directory",strlen("cd: no such file or directory"));
+        char error[256];
+        snprintf(error, sizeof(error), "bash: cd: %s: No such file or directory\n", dirName);
+        write(1,error,strlen(error));
     }
 }
 
 
 void copyFile(char *sourcePath, char *destinationPath){
+    struct stat checker;
+    if(stat(sourcePath, &checker) != 0){
+        char error[256];
+        snprintf(error, sizeof(error), "cp: cannot stat '%s\n': No such file or directory", sourcePath);
+        write(1,error,strlen(error));
+        return;
+    }
+
     struct stat path_stat;
-    stat(destinationPath,&path_stat);
+    int result = stat(destinationPath,&path_stat);
+    if (result == -1){
+        char error[256];
+        snprintf(error, sizeof(error), "cp: cannot create regular file '%s\n': No such file or directory", destinationPath);
+        write(1,error,strlen(error));
+    }
     char* filename = basename(sourcePath);
     int dest;
 
@@ -65,7 +75,6 @@ void copyFile(char *sourcePath, char *destinationPath){
         dest = open(destinationPath, O_WRONLY | O_CREAT | O_TRUNC, 0644);
     }
     
-
     int src = open(sourcePath, O_RDONLY);
     char buffer[4096];
     ssize_t bytes;
@@ -75,12 +84,26 @@ void copyFile(char *sourcePath, char *destinationPath){
 
     close(src);
     close(dest);
-
     if (S_ISDIR(path_stat.st_mode)){free(buffer);}    
 }
 
 
 void moveFile(char *sourcePath, char *destinationPath){
+    struct stat checker1;
+    if(stat(sourcePath, &checker1) != 0){
+        char error[256];
+        snprintf(error, sizeof(error), "mv: cannot stat '%s\n': No such file or directory", sourcePath);
+        write(1,error,strlen(error));
+        return;
+    }
+    struct stat checker2;
+    if(stat(destinationPath, &checker2) != 0){
+        char error[256];
+        snprintf(error, sizeof(error), "mv: cannot stat '%s\n': No such file or directory", destinationPath);
+        write(1,error,strlen(error));
+        return;
+    }
+
     copyFile(sourcePath,destinationPath);
     deleteFile(sourcePath);
 }
@@ -100,7 +123,9 @@ void deleteFile(char *filename){
     }
 
     if(found == 0){
-        write(1,"No such file or directory",strlen("No such file or directory"));
+        char error[256];
+        snprintf(error, sizeof(error), "rm: cannot remove ‘%s\n’: No such file or directory", filename);
+        write(1,error,strlen(error));
     }
     else{
         remove(filename);
@@ -109,6 +134,14 @@ void deleteFile(char *filename){
 
 
 void displayFile(char *filename){
+    struct stat checker;
+    if(stat(filename, &checker) != 0){
+        char error[256];
+        snprintf(error, sizeof(error), "cat: '%s\n': No such file or directory", filename);
+        write(1,error,strlen(error));
+        return;
+    }
+
     int src = open("input.txt", O_RDONLY);
     char buffer[4096];
     ssize_t bytes;
